@@ -2,8 +2,8 @@ package trie
 
 // Node is the trie's node
 type Node struct {
-	Data     interface{}
 	children map[string]*Node
+	Data     any
 }
 
 func (n *Node) getChild(s string) *Node {
@@ -18,7 +18,7 @@ func (n *Node) addChild(s string, child *Node) {
 	n.children[s] = child
 }
 
-func newNode(data interface{}) *Node {
+func newNode(data any) *Node {
 	return &Node{
 		Data:     data,
 		children: map[string]*Node{},
