@@ -3,8 +3,9 @@ package dialer
 import "go.uber.org/atomic"
 
 var (
-	DefaultOptions   []Option
-	DefaultInterface = atomic.NewString("")
+	DefaultOptions     []Option
+	DefaultInterface   = atomic.NewString("")
+	DefaultRoutingMark = atomic.NewInt32(0)
 )
 
 type option struct {
