@@ -1,21 +1,16 @@
 package rules
 
 import (
-	"fmt"
-	"strconv"
+	"path/filepath"
 	"strings"
 
-	"github.com/Dreamacro/clash/common/cache"
-	"github.com/Dreamacro/clash/component/process"
 	C "github.com/Dreamacro/clash/constant"
-	"github.com/Dreamacro/clash/log"
 )
 
-var processCache = cache.NewLRUCache(cache.WithAge(2), cache.WithSize(64))
-
 type Process struct {
-	adapter string
-	process string
+	adapter  string
+	process  string
+	nameOnly bool
 }
 
 func (ps *Process) RuleType() C.RuleType {
@@ -23,26 +18,11 @@ func (ps *Process) RuleType() C.RuleType {
 }
 
 func (ps *Process) Match(metadata *C.Metadata) bool {
-	key := fmt.Sprintf("%s:%s:%s", metadata.NetWork.String(), metadata.SrcIP.String(), metadata.SrcPort)
-	cached, hit := processCache.Get(key)
-	if !hit {
-		srcPort, err := strconv.Atoi(metadata.SrcPort)
-		if err != nil {
-			processCache.Set(key, "")
-			return false
-		}
-
-		name, err := process.FindProcessName(metadata.NetWork.String(), metadata.SrcIP, srcPort)
-		if err != nil {
-			log.Debugln("[Rule] find process name %s error: %s", C.Process.String(), err.Error())
-		}
-
-		processCache.Set(key, name)
-
-		cached = name
+	if ps.nameOnly {
+		return strings.EqualFold(filepath.Base(metadata.ProcessPath), ps.process)
 	}
 
-	return strings.EqualFold(cached.(string), ps.process)
+	return strings.EqualFold(metadata.ProcessPath, ps.process)
 }
 
 func (ps *Process) Adapter() string {
@@ -57,9 +37,14 @@ func (ps *Process) ShouldResolveIP() bool {
 	return false
 }
 
-func NewProcess(process string, adapter string) (*Process, error) {
+func (ps *Process) ShouldFindProcess() bool {
+	return true
+}
+
+func NewProcess(process string, adapter string, nameOnly bool) (*Process, error) {
 	return &Process{
-		adapter: adapter,
-		process: process,
+		adapter:  adapter,
+		process:  process,
+		nameOnly: nameOnly,
 	}, nil
 }
